initial _train method
